+++ src/main/java/microservice/cloud/inventory/category/application/use_cases/DeleteCategoryUseCase.java
@@ -4,7 +4,8 @@ import microservice.cloud.inventory.category.application.ports.in.DeleteCategory
 import microservice.cloud.inventory.category.domain.entity.Category;
 import microservice.cloud.inventory.category.domain.repository.CategoryRepository;
 import microservice.cloud.inventory.shared.application.ports.out.GetMePort;
-import microservice.cloud.inventory.shared.domain.value_objects.Id;
+import microservice.cloud.inventory.shared.domain.value_objects.Me;
+import microservice.cloud.inventory.shared.domain.value_objects.Permission;
 import microservice.cloud.inventory.shared.domain.value_objects.Slug;
 
 public class DeleteCategoryUseCase implements DeleteCategoryUseCasePort {
@@ -22,9 +23,14 @@ public class DeleteCategoryUseCase implements DeleteCategoryUseCasePort {
 
     @Override
     public void execute(Slug find_slug) {
-        Category category = categoryRepository.findBySlug(find_slug);
+        Me me = getMePort.execute();
 
-        category.canIDeleteThisCategory(getMePort.execute());
+        if(me == null)
+            throw new RuntimeException("You do not have permission to perform this action.");
+
+        me.IHavePermission(Permission.deleteCategory());
+
+        Category category = categoryRepository.findBySlug(find_slug);
 
         categoryRepository.delete(category);
     }
