+++ src/main/java/microservice/cloud/inventory/product/application/use_cases/DeleteProductUseCase.java
@@ -1,9 +1,11 @@
 package microservice.cloud.inventory.product.application.use_cases;
 
 import microservice.cloud.inventory.product.application.ports.in.DeleteProductUseCasePort;
-import microservice.cloud.inventory.shared.application.ports.out.GetMePort;
 import microservice.cloud.inventory.product.domain.entity.Product;
 import microservice.cloud.inventory.product.domain.entity.ProductRepository;
+import microservice.cloud.inventory.shared.application.ports.out.GetMePort;
+import microservice.cloud.inventory.shared.domain.value_objects.Me;
+import microservice.cloud.inventory.shared.domain.value_objects.Permission;
 import microservice.cloud.inventory.shared.domain.value_objects.Slug;
 
 public class DeleteProductUseCase implements DeleteProductUseCasePort {
@@ -21,10 +23,15 @@ public class DeleteProductUseCase implements DeleteProductUseCasePort {
 
     @Override
     public void execute(Slug find_slug) {
+        Me me = getMePort.execute();
+        
+        if(me == null)
+            throw new RuntimeException("You do not have permission to perform this action");
+
+        me.IHavePermission(Permission.deleteProduct());
+
         Product product = productRepository.findBySlug(find_slug);
-
-        Product.canIDeleteThisProduct(getMePort.execute());
-
+        
         productRepository.delete(product);
     }
 }
